@@ -1,21 +1,29 @@
+/**
+ * plugins/vuetify.js
+ *
+ * Framework documentation: https://vuetifyjs.com`
+ */
+
 // Styles
 import "@mdi/font/css/materialdesignicons.css";
 import "vuetify/styles";
 
-// Vuetify
+// Composables
 import { createVuetify } from "vuetify";
 
+// https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
 export default createVuetify({
   theme: {
     defaultTheme: "dark",
     variations: {
-      colors: ["primary", "secondary"],
-      lighten: 1,
-      darken: 2,
+      colors: ["primary", "secondary", "color-1", "color-2", "color-3"],
+      lighten: 10,
+      darken: 5,
     },
     themes: {
       light: {
         colors: {
+          background: "#EAEDFF",
           surface: "#F5F5F5",
           primary: "#2A3E72",
           secondary: "#0DCCC2",
@@ -24,36 +32,22 @@ export default createVuetify({
           info: "#2196F3",
           success: "#4CAF50",
           warning: "#FB8C00",
-
-          "fill-1": "#434656",
-          "fill-2": "#434656",
-          "fill-3": "#0DCCC2",
-          "stroke-1": "#434656",
-          "stroke-12": "#FFFFFF",
-          "stroke-2": "#3CC47C", //"#F85C5C",
-          "stroke-3": "#F3B24A",
-          "stroke-4": "#1C222B",
         },
       },
       dark: {
         colors: {
-          surface: "#1C222B",
-          primary: "#00fffa",
+          background: "#000109",
+          surface: "#000109",
+          primary: "#EAEDFF",
           secondary: "#03DAC6",
           heading: "#042070",
           error: "#B00020",
           info: "#00fffa",
           success: "#4CAF50",
           warning: "#FB8C00",
-
-          "fill-1": "#FFFFFF",
-          "fill-2": "#434656",
-          "fill-3": "#0DCCC2",
-          "stroke-1": "#FFFFFF",
-          "stroke-12": "#FFFFFF",
-          "stroke-2": "#3CC47C", //"#F85C5C",
-          "stroke-3": "#F3B24A",
-          "stroke-4": "#F5F5F5",
+          "color-1": "#3CC47C", //green
+          "color-2": "#44C0EF", //blue
+          "color-3": "#7F56D9", //purple
         },
       },
     },
