@@ -1,31 +1,36 @@
-import {createRouter, createWebHistory} from 'vue-router'
 
-const routes = [
-  {
-    path: '/',
-    name: 'home',
-    meta: {title: "Home"},
-    component: () => import(/* webpackChunkName: "about" */ '../views/HomePage')
-  }
-]
+/**
+ * router/index.ts
+ *
+ * Automatic routes for `./src/pages/*.vue`
+ */
+
+// Composables
+import { createRouter, createWebHistory } from 'vue-router/auto'
+import { routes } from 'vue-router/auto-routes'
 
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes,
-  scrollBehavior() {
-    // always scroll to top
-    return {top: 0}
-  },
 })
 
+// Workaround for https://github.com/vitejs/vite/issues/11804
+router.onError((err, to) => {
+  if (err?.message?.includes?.('Failed to fetch dynamically imported module')) {
+    if (!localStorage.getItem('vuetify:dynamic-reload')) {
+      console.log('Reloading page to fix dynamic import error')
+      localStorage.setItem('vuetify:dynamic-reload', 'true')
+      location.assign(to.fullPath)
+    } else {
+      console.error('Dynamic import error, reloading page did not fix it', err)
+    }
+  } else {
+    console.error(err)
+  }
+})
 
-/* redirecting to home page if access token is not provided*/
-router.beforeEach((to, from, next) => {
-  next()
-  //setting the title
-  if (to && to.meta && to.meta.title)
-    window.document.title = "Amit Kumar Sahu | " + to.meta.title;
-});
-
+router.isReady().then(() => {
+  localStorage.removeItem('vuetify:dynamic-reload')
+})
 
 export default router
